backend/requirements.txt
 flask
 flask-mysql
 mysql-connector-python
 werkzeug 
 flask-sqlalchemy 
 flask-cors 
 flask-socketio 
